ml-service/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Machine Learning Libraries
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.24.4
scipy==1.11.4

# Database connections
pymongo==4.6.0
redis==5.0.1

# Data processing
joblib==1.3.2
pickle-mixin==1.0.2

# HTTP requests
httpx==0.25.2
aiohttp==3.9.1

# Logging and monitoring
python-json-logger==2.0.7
prometheus-client==0.19.0

# Development dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0

# Optional: For advanced ML features
surprise==1.1.3  # For collaborative filtering
implicit==0.7.2  # For matrix factorization